feat: Modify diction test API response to return integer (#35)

+++ Server/src/main/java/com/pro_diction/server/domain/test/dto/TestResponseDto.java
@@ -9,5 +9,5 @@ import lombok.*;
 @NoArgsConstructor
 public class TestResponseDto {
     private Long studyId;
-    private Double score;
+    private Integer score;
 }
